requirements.txt
attrs==24.2.0
blinker==1.4
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
exceptiongroup==1.2.2
google-api-core==2.19.2
google-api-python-client==2.144.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httplib2==0.22.0
hyperframe==6.0.1
idna==3.8
kaitaistruct==0.10
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.2.1
pyparsing==3.1.4
PySocks==1.7.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.24.0
selenium-wire==5.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
temp==2020.7.2
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
zstandard==0.23.0
setuptools==75.2.0
